bus/hp_dio/98603a: convert to anonymous namespace (nw)
